social-backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Daniil Bibik"]
license = "GNU-GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = {extras = ["argon2"], version = "^4.1.4"}
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
uwsgi = "^2.0.21"
python-decouple = "^3.6"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}
django-redis-cache = "^3.0.1"
# HiRedis parser for Django Redis cache
hiredis = "^2.1.0"
drf-orjson-renderer = "^1.7.1"
serpy = "^0.3.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"
behave-django = "^1.4.0"
mixer = "^7.2.2"
coverage = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
